requirements.txt
alpaca-py
cvxpy
pandas
scikit-learn
numpy